Creating component Dokumentasi, adding grup wa link, etc
