+++ WebCommunity/package.json
@@ -18,6 +18,7 @@
     "react-dom": "^18.3.1",
     "react-icons": "^5.5.0",
     "react-router-dom": "^7.3.0",
+    "react-scroll": "^1.9.3",
     "tailwindcss": "^4.0.12"
   },
   "devDependencies": {
