+++ WebCommunity/src/App.jsx
@@ -3,14 +3,14 @@ import { BrowserRouter as Router, Routes, Route} from 'react-router-dom'
 import Beranda from './components/beranda'
 import Teknologi from './components/pageTekno'
 import Olahraga from './components/pageOlahraga'
-import Webdev from './components/Webdev/webdevPage'
-import OSC from './components/OSC/PageOSC'
-import GameKita from './components/GameKita/PageGamekita'
-import ShuttleBlue from './components/ShuttleBlue/PageShuttleBlue'
-import BlueIce from './components/BlueIce/PageBlueIce'
-import BlueKnight from './components/BlueKnight/PageBlueKnight'
-import IBall from './components/I-Ball/PageIBall'
-import Isport from './components/I-Sport/PageISport'
+import Webdev from './components/Teknologi/webdevPage'
+import OSC from './components/Teknologi/PageOSC'
+import GameKita from './components/Teknologi/PageGamekita'
+import ShuttleBlue from './components/Olahraga/PageShuttleBlue'
+import BlueIce from './components/Olahraga/PageBlueIce'
+import BlueKnight from './components/Olahraga/PageBlueKnight'
+import IBall from './components/Olahraga/PageIBall'
+import Isport from './components/Olahraga/PageISport'
 import Event from './components/Event'
 
 function App() {
