+++ WebCommunity/src/components/Olahraga/PageIBall.jsx
@@ -1,31 +1,36 @@
 import '../../App.css'
 import { Link } from 'react-router-dom'
+import Navbar from '../navbar'
+import Footer from "../footer"
 import FadeContent from '../../AwesomeStyle/FadeContent/FadeContent'
 import AnimatedContent from '../../AwesomeStyle/AnimatedContent/AnimatedContent'
-import Navbar from '../navbar'
-import shuttle from '../../assets/shuttle.jpeg'
-import Footer from '../footer'
+import ball1 from '../../assets/Iball1.jpg'
+import ball2 from '../../assets/Iball2.jpeg'
+import ball3 from '../../assets/Iball3.jpeg'
 import { useEffect } from 'react'
+import Dokumentasi from '../Dokumentasi'
 
-export default function ShuttleBlue() {
+const images =[ball1, ball2, ball3]
+
+export default function IBall() {
     useEffect(() => {
         window.scrollTo(0, 0);
       }, []);
-    return(
+    return (
         <div>
-            <Navbar/>
-            
+            <Navbar />
+
             <div>
                 <div className='px-8 md:px-[10rem] mt-28'>
                     <Link to="/KomunitasOlahraga" className="font-poppins font-semibold underline text-blue-900">
                         Kembali ke halaman komunitas
                     </Link>
                 </div>
-                
+
                 <div className="flex flex-col md:flex-row justify-center gap-4 lg:gap-16 items-center mt-9 px-6">
                     <FadeContent blur={true} duration={1000} easing="ease-out" initialOpacity={0}>
                         <div>
-                            <img src={shuttle} alt="Web Developer" className="w-[22rem] h-[22rem] object-contain" />
+                            <img src={ball2} alt="I-Ball" className="w-[22rem] h-[22rem] object-cover rounded-lg" />
                         </div>
                     </FadeContent>
 
@@ -33,23 +38,23 @@ export default function ShuttleBlue() {
                         <div className="max-w-md text-center md:text-left">
                             <div className="mb-6 space-y-3">
                                 <h1 className="text-4xl font-poppins text-[#174B86] font-bold">
-                                    Shuttle Blue
-                                </h1>   
+                                    I - Ball
+                                </h1>
                                 <p className="text-justify text-gray-700">
-                                Shuttle Blue adalah komunitas bagi mahasiswa yang memiliki 
-                                ketertarikan dalam olahraga badminton. Komunitas ini 
-                                menyediakan ruang bagi para pemain dari berbagai level 
-                                untuk berlatih, bertanding, dan meningkatkan keterampilan 
-                                dalam suasana yang kompetitif namun tetap menyenangkan.
+                                I - Ball merupakan komunitas olahraga Basket yang menjadi tempat bagi mahasiswa 
+                                untuk mengasah keterampilan, meningkatkan teamwork, serta berpartisipasi dalam 
+                                pertandingan seru-seru an dan kegiatan basket lainnya.
                                 </p>
                             </div>
 
                             <div className="space-y-3">
-                                <h3 className="text-lg font-bold text-[#174B86]">Bidang Olahraga:</h3>
+                                <h3 className="text-lg font-bold text-[#174B86]">Topik Pembahasan:</h3>
                                 <div className="flex flex-wrap justify-center md:justify-start gap-3">
-                                    <span className="text-sm py-2 px-4 bg-[#1A57A7] rounded-full text-white">
-                                        Permainan Badminton
-                                    </span>
+                                    {["Permainan Bola Basket"].map(topic => (
+                                        <span key={topic} className="text-sm py-2 px-4 bg-[#1A57A7] rounded-full text-white">
+                                            {topic}
+                                        </span>
+                                    ))}
                                 </div>
                             </div>
                         </div>
@@ -65,15 +70,17 @@ export default function ShuttleBlue() {
                     animateOpacity
                     scale={1}
                     threshold={0.2}
-                    >
+                >
                     <div className="flex justify-center mt-8 md:mt-10">
-                        <Link to="/" className="py-3 px-6 bg-emerald-500 rounded-full text-white font-semibold shadow-md hover:bg-emerald-600 transition">
+                        <Link to="https://www.google.com/url?q=https://chat.whatsapp.com/K4gHoZhpPfF6lqpfbOtWPf&sa=D&source=editors&ust=1742044961512052&usg=AOvVaw2Ygq_R4Q84TaEg3TpbaUvJ" target='_blank' className="py-3 px-6 bg-emerald-500 rounded-full text-white font-semibold shadow-md hover:bg-emerald-600 transition">
                             Gabung dengan komunitas
                         </Link>
                     </div>
                 </AnimatedContent>
             </div>
-            <Footer/>
+
+            <Dokumentasi dokum={images}/>
+            <Footer />
         </div>
     )
 }
