+++ WebCommunity/src/components/Olahraga/PageShuttleBlue.jsx
@@ -1,14 +1,13 @@
 import '../../App.css'
-import Navbar from '../navbar'
-import webdev from '../../assets/webdev.jpeg'
 import { Link } from 'react-router-dom'
-import WebdevDocum from './webdevDocum'
-import Footer from '../footer'
 import FadeContent from '../../AwesomeStyle/FadeContent/FadeContent'
 import AnimatedContent from '../../AwesomeStyle/AnimatedContent/AnimatedContent'
+import Navbar from '../navbar'
+import shuttle from '../../assets/shuttle.jpeg'
+import Footer from '../footer'
 import { useEffect } from 'react'
 
-export default function Webdev() {
+export default function ShuttleBlue() {
     useEffect(() => {
         window.scrollTo(0, 0);
       }, []);
@@ -18,7 +17,7 @@ export default function Webdev() {
             
             <div>
                 <div className='px-8 md:px-[10rem] mt-28'>
-                    <Link to="/KomunitasTeknologi" className="font-poppins font-semibold underline text-blue-900">
+                    <Link to="/KomunitasOlahraga" className="font-poppins font-semibold underline text-blue-900">
                         Kembali ke halaman komunitas
                     </Link>
                 </div>
@@ -26,7 +25,7 @@ export default function Webdev() {
                 <div className="flex flex-col md:flex-row justify-center gap-4 lg:gap-16 items-center mt-9 px-6">
                     <FadeContent blur={true} duration={1000} easing="ease-out" initialOpacity={0}>
                         <div>
-                            <img src={webdev} alt="Web Developer" className="w-[22rem] h-[22rem] object-contain" />
+                            <img src={shuttle} alt="Web Developer" className="w-[22rem] h-[22rem] object-contain" />
                         </div>
                     </FadeContent>
 
@@ -34,25 +33,23 @@ export default function Webdev() {
                         <div className="max-w-md text-center md:text-left">
                             <div className="mb-6 space-y-3">
                                 <h1 className="text-4xl font-poppins text-[#174B86] font-bold">
-                                    Web Developer
-                                </h1>
+                                    Shuttle Blue
+                                </h1>   
                                 <p className="text-justify text-gray-700">
-                                    Wadah pengembangan bagi mahasiswa yang tertarik dan ingin berkembang 
-                                    di dunia pengembangan web. Di sini, kamu bisa belajar, berbagi, dan 
-                                    berkolaborasi dalam menciptakan website yang fungsional dan menarik. 
-                                    Komunitas ini juga menjadi wadah untuk mengeksplorasi tren terbaru dan 
-                                    mengasah kreativitas.
+                                Shuttle Blue adalah komunitas bagi mahasiswa yang memiliki 
+                                ketertarikan dalam olahraga badminton. Komunitas ini 
+                                menyediakan ruang bagi para pemain dari berbagai level 
+                                untuk berlatih, bertanding, dan meningkatkan keterampilan 
+                                dalam suasana yang kompetitif namun tetap menyenangkan.
                                 </p>
                             </div>
 
                             <div className="space-y-3">
-                                <h3 className="text-lg font-bold text-[#174B86]">Topik Pembahasan:</h3>
+                                <h3 className="text-lg font-bold text-[#174B86]">Bidang Olahraga:</h3>
                                 <div className="flex flex-wrap justify-center md:justify-start gap-3">
-                                    {["Pemrograman Website", "Node JS", "Tailwind CSS"].map(topic => (
-                                            <span key={topic} className="text-sm py-2 px-4 bg-[#1A57A7] rounded-full text-white">
-                                                {topic}
-                                            </span>
-                                    ))}
+                                    <span className="text-sm py-2 px-4 bg-[#1A57A7] rounded-full text-white">
+                                        Permainan Badminton
+                                    </span>
                                 </div>
                             </div>
                         </div>
@@ -70,14 +67,12 @@ export default function Webdev() {
                     threshold={0.2}
                     >
                     <div className="flex justify-center mt-8 md:mt-10">
-                        <Link to="/" className="py-3 px-6 bg-emerald-500 rounded-full text-white font-semibold shadow-md hover:bg-emerald-600 transition">
+                        <Link to="https://www.google.com/url?q=https://chat.whatsapp.com/DhlFISdH4nvCGzmfSwYLst&sa=D&source=editors&ust=1742045009661874&usg=AOvVaw0NF759cjwIMYON7oA4rPcA" target='_blank' className="py-3 px-6 bg-emerald-500 rounded-full text-white font-semibold shadow-md hover:bg-emerald-600 transition">
                             Gabung dengan komunitas
                         </Link>
                     </div>
                 </AnimatedContent>
             </div>
-            
-            <WebdevDocum/>
             <Footer/>
         </div>
     )
