+++ WebCommunity/src/components/Teknologi/webdevPage.jsx
@@ -1,32 +1,38 @@
 import '../../App.css'
-import { Link } from 'react-router-dom'
 import Navbar from '../navbar'
-import Footer from "../footer"
+import webdev from '../../assets/webdev.jpeg'
+import { Link } from 'react-router-dom'
+import Footer from '../footer'
 import FadeContent from '../../AwesomeStyle/FadeContent/FadeContent'
 import AnimatedContent from '../../AwesomeStyle/AnimatedContent/AnimatedContent'
-import BlueKnek from '../../assets/BlueKnight.jpeg'
-import DocumBlueKnight from './DocumBlueKnight'
+import webdev1 from '../../assets/webdev1.jpeg'
+import webdev2 from '../../assets/webdev2.jpeg'
+import webdev3 from '../../assets/webdev3.jpeg'
+import webdev4 from '../../assets/webdev4.jpeg'
 import { useEffect } from 'react'
+import Dokumentasi from '../Dokumentasi'
 
-export default function BlueKnight() {
+const images = [webdev1, webdev2, webdev3, webdev4];
+
+export default function Webdev() {
     useEffect(() => {
         window.scrollTo(0, 0);
       }, []);
-    return (
+    return(
         <div>
-            <Navbar />
-
+            <Navbar/>
+            
             <div>
                 <div className='px-8 md:px-[10rem] mt-28'>
-                    <Link to="/KomunitasOlahraga" className="font-poppins font-semibold underline text-blue-900">
+                    <Link to="/KomunitasTeknologi" className="font-poppins font-semibold underline text-blue-900">
                         Kembali ke halaman komunitas
                     </Link>
                 </div>
-
+                
                 <div className="flex flex-col md:flex-row justify-center gap-4 lg:gap-16 items-center mt-9 px-6">
                     <FadeContent blur={true} duration={1000} easing="ease-out" initialOpacity={0}>
                         <div>
-                            <img src={BlueKnek} alt="Web Developer" className="w-[22rem] h-[22rem] object-contain" />
+                            <img src={webdev} alt="Web Developer" className="w-[22rem] h-[22rem] object-contain" />
                         </div>
                     </FadeContent>
 
@@ -34,21 +40,24 @@ export default function BlueKnight() {
                         <div className="max-w-md text-center md:text-left">
                             <div className="mb-6 space-y-3">
                                 <h1 className="text-4xl font-poppins text-[#174B86] font-bold">
-                                    Blue Knight
+                                    Web Developer
                                 </h1>
                                 <p className="text-justify text-gray-700">
-                                Blue Knight merupakan komunitas olahraga voli yang menjadi tempat bagi mahasiswa untuk mengasah keterampilan, 
-                                meningkatkan teamwork, serta berpartisipasi dalam pertandingan seru-seru an dan kegiatan voli lainnya.
+                                    Wadah pengembangan bagi mahasiswa yang tertarik dan ingin berkembang 
+                                    di dunia pengembangan web. Di sini, kamu bisa belajar, berbagi, dan 
+                                    berkolaborasi dalam menciptakan website yang fungsional dan menarik. 
+                                    Komunitas ini juga menjadi wadah untuk mengeksplorasi tren terbaru dan 
+                                    mengasah kreativitas.
                                 </p>
                             </div>
 
                             <div className="space-y-3">
                                 <h3 className="text-lg font-bold text-[#174B86]">Topik Pembahasan:</h3>
                                 <div className="flex flex-wrap justify-center md:justify-start gap-3">
-                                    {["Permainan Bola Volly"].map(topic => (
-                                        <span key={topic} className="text-sm py-2 px-4 bg-[#1A57A7] rounded-full text-white">
-                                            {topic}
-                                        </span>
+                                    {["Pemrograman Website", "Node JS", "Tailwind CSS"].map(topic => (
+                                            <span key={topic} className="text-sm py-2 px-4 bg-[#1A57A7] rounded-full text-white">
+                                                {topic}
+                                            </span>
                                     ))}
                                 </div>
                             </div>
@@ -65,17 +74,17 @@ export default function BlueKnight() {
                     animateOpacity
                     scale={1}
                     threshold={0.2}
-                >
+                    >
                     <div className="flex justify-center mt-8 md:mt-10">
-                        <Link to="/" className="py-3 px-6 bg-emerald-500 rounded-full text-white font-semibold shadow-md hover:bg-emerald-600 transition">
+                        <Link to="https://chat.whatsapp.com/Fv51UKt6IhPHIXA5ZP5hae" target='_blank' className="py-3 px-6 bg-emerald-500 rounded-full text-white font-semibold shadow-md hover:bg-emerald-600 transition">
                             Gabung dengan komunitas
                         </Link>
                     </div>
                 </AnimatedContent>
             </div>
-
-            <DocumBlueKnight />
-            <Footer />
+            
+            <Dokumentasi dokum={images}/>
+            <Footer/>
         </div>
     )
 }
