+++ WebCommunity/src/components/beranda.jsx
@@ -1,13 +1,24 @@
 import '../App.css'
 import Navbar from './navbar'
-import { Link } from 'react-router-dom'
 import Manfaat from './HomePage/benefit'
 import Overview from './HomePage/overview'
 import Explore from './HomePage/explore'
 import Footer from './footer'
 import AnimatedContent from '../AwesomeStyle/AnimatedContent/AnimatedContent'
+import { useNavigate } from 'react-router-dom'
+import { useEffect } from 'react'
 
 export default function Beranda() {
+    useEffect(() => {
+            window.scrollTo(0, 0);
+    }, [])
+
+    const navigate = useNavigate()
+
+    const handleScroll = () => {
+        navigate('/#Jelajahi') 
+    }
+
     return (
         <div>
             <Navbar/>
@@ -23,7 +34,7 @@ export default function Beranda() {
                     threshold={0.2}
                     >
                     <div>
-                        <h1 className='text-4xl md:text-6xl font-poppins text-[#174B86] font-bold leading-tight md:leading-16'>
+                        <h1 className='text-4xl md:text-6xl font-poppins text-[#174B86] font-bold leading-tight md:leading-18'>
                         Wadah Pengembangan,<br/>Kolaborasi, dan Inovasi HIMIT</h1>
                     </div>
                 </AnimatedContent>
@@ -41,10 +52,10 @@ export default function Beranda() {
                     <div>
                         <p className='font-poppins font-light text-center text-base md:text-lg mt-6 max-w-lg md:max-w-2xl'>
                             HIMIT menghadirkan berbagai komunitas untuk mendukung pengembangan keterampilan, 
-                            inovasi, dan kolaborasi mahasiswa Teknik Informatika. Temukan komunitas yang 
+                            inovasi, dan kolaborasi mahasiswa Teknik Informatika dan Sains Data. Temukan komunitas yang 
                             sesuai dengan minatmu dan jadilah bagian dari perubahan!</p>
-                        <button className='mt-4 py-3 px-5 bg-[#1A57A7] rounded-full font-poppins text-white font-semibold text-sm md:text-base transition duration-300 ease-in-out hover:translate-y-[-4px] hover:shadow-lg'>
-                            <Link to='/'>Jelajahi Komunitas</Link>
+                        <button onClick={handleScroll} className='mt-4 py-3 px-5 bg-[#1A57A7] rounded-full font-poppins text-white font-semibold text-sm md:text-base transition duration-300 ease-in-out hover:translate-y-[-4px] hover:shadow-lg'>
+                            Jelajahi Komunitas
                         </button>
                     </div>
                 </AnimatedContent>
